backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
sentence-transformers>=2.2.2
chromadb>=0.4.18
rank-bm25>=0.2.2
numpy>=1.26.0
pandas>=2.1.3
pypdf>=3.17.1
python-docx>=1.1.0
tqdm>=4.66.1
python-dotenv>=1.0.0
google-generativeai>=0.3.0
# OCR support for scanned PDFs and presentations
PyMuPDF>=1.23.0  # fitz - better PDF handling and OCR
pytesseract>=0.3.10  # OCR engine wrapper
pdf2image>=1.16.3  # Convert PDF pages to images for OCR
Pillow>=10.0.0  # Image processing for OCR
# LlamaIndex for better LLM reliability (minimal install)
llama-index-core>=0.10.0
llama-index-llms-gemini>=0.1.0

